troquei de inteiros para doubles e comecei a implementar as apostas :)
Refatoração de tipos e lógica para uso de double

Renomeação e ajuste de funções em Fichas, Jogador e Mesa.
- `Fichas::veResultado` renomeada para `Fichas::resultadojogador` e tipo de retorno alterado para `double`.
- `Fichas::compraFichas`, `Fichas::fichasIniciais` e `Fichas::aposta` agora aceitam `double& saldoJogador` e retornam `double`.
- Variáveis `fichasJogador` e `fichasMesa` em `Fichas` alteradas para `double`.
- Funções `Jogador::calculaPontos`, `Jogador::retornaSaldoFichas` e `Mesa::calculaPontosMesa`, `Mesa::retornaSaldoFichasMesa` alteradas para `double`.
- Adicionadas funções `Jogador::valorAposta` e `Jogador::condicaoDeVitoria`.
- Variável `saldoFichas` em `Jogador` alterada para `double`.
- Ajustes no `main.cpp` para refletir mudanças nas funções e tipos.
- Remoção de comentários e código desnecessário.

+++ projeto21/mesa.cpp
@@ -42,8 +42,8 @@ void Mesa::mostratudo(){
 	}
 }
 
-int Mesa::calculaPontosMesa() {
-	int total = 0;
+double Mesa::calculaPontosMesa() {
+	double total = 0;
 	for (const auto& cartas : cartaMesa) {
 		if (cartas.numero_Carta == 1) {
 			total += 11;
@@ -63,7 +63,7 @@ int Mesa::calculaPontosMesa() {
 	return total;
 }
 
-int Mesa::retornaSaldoFichasMesa() {
+double Mesa::retornaSaldoFichasMesa() {
 	return fichasMesa;
 }
 //if (jogador) {
